:rocket: feat(issue): rename 및 오류 해결 #13
