+++ src/pages/store/StoreDetail.js
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import StoreApi from '../../apis/StoreApi';
 import Map from './Sections/Map';
-import { Review } from './Sections/Review';
+import { Reviews } from './Sections/Reviews';
 import { useCookies } from 'react-cookie';
 
 import { Card, Avatar, Col, Typography, Row, List } from 'antd';
@@ -29,7 +29,7 @@ const StoreDetail = () => {
 			<div style={{ width: '85%', margin: '3rem auto' }}>
 				<Title level={4}> Store info </Title>
 				<hr />
-				<Review
+				<Reviews
 					refreshFunction={getStoreData}
 					newReview={_store.newReview}
 					ReviewLists={_store.reviews}
